Removed kernal checks
I've removed kernel checks from install.sh, with the offical ubuntu
image the kernal that is running is the only one we need to install for.
This means we don't need to check for other kernels and we can remove
`-k` from the dkms invocations. (When you use dkms if `-k` is ommitted
then it assumes the currentluy running kernel.)